pyproject.toml
[project]
name = "data-cleaning-env"
version = "0.1.0"
description = "OpenEnv environment for AI-powered data cleaning"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
    "polars>=1.0",
    "openpyxl>=3.1",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "duckdb>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
]
inference = [
    "openai>=1.0",
]
ui = [
    "gradio>=4.0",
    "plotly>=5.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
